pyproject.toml
[project]
name = "shopbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram==3.18.0",
    "colorama==0.4.6",
    "python-dotenv==1.0.1",
    "sqlalchemy>=2.0.40",
]
